requirements.txt
Django==5.0.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==25.1
gunicorn==25.1.0
